requirements.txt
absl-py==1.4.0
addict==2.4.0
aggdraw==1.3.15
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.2.1
anyio==4.2.0
argcomplete==3.2.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.3
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bezier==2021.2.12
bleach==5.0.1
blinker==1.6.2
boto3==1.34.35
botocore==1.34.35
branca==0.5.0
Brotli==1.1.0
brotlipy==0.7.0
cachetools==5.3.0
cffi==1.15.1
chardet==3.0.4
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
crowdposetools==2.0
cycler==0.11.0
Cython==3.0.8
dacite==1.7.0
datasets==2.17.1
debugpy==1.6.2
decorator==5.1.1
deformable-attention==0.0.18
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.23.1
dill==0.3.8
dnspython==2.5.0
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.6.0
einops==0.7.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==0.8.3
faiss==1.7.2
fast-slic==0.4.0
fastjsonschema==2.15.3
fiftyone==0.23.4
fiftyone-brain==0.16.0
fiftyone-db==1.1.1
filelock==3.13.1
Fiona==1.8.21
Flask==2.3.3
folium==0.12.1.post1
fonttools==4.34.4
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
fugashi==1.1.2
future==0.18.2
fvcore==0.1.5.post20221221
geographiclib==1.52
geopandas==0.11.1
geopy==2.2.0
gitdb==4.0.9
GitPython==3.1.29
glob2==0.7
google-auth==2.16.1
google-auth-oauthlib==0.4.6
googletrans==4.0.0rc1
graphql-core==3.2.3
grpcio==1.51.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.9.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.19.4
humanize==4.9.0
Hypercorn==0.16.0
hyperframe==5.2.0
idna==2.10
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==7.0.1
importlib-resources==5.8.0
inflate64==1.0.0
intel-openmp==2022.1.0
iopath==0.1.10
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jamo==0.4.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonlines==4.0.0
jsonschema==4.6.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kaleido==0.2.1
kiwisolver==1.4.3
kmeans-pytorch==0.3
lightning-utilities==0.10.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mkl==2022.1.0
mongoengine==0.24.2
motor==3.3.2
multidict==6.0.5
multiprocess==0.70.16
MultiScaleDeformableAttention==1.0
multivolumefile==0.2.3
munch==2.5.0
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
notebook==6.4.12
numpy==1.23.1
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plac==1.3.5
platformdirs==4.2.0
plotly==5.18.0
portalocker==2.8.2
pprintpp==0.4.0
priority==2.0.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.20.1
psd-tools==1.9.21
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-hanspell==1.1
py7zr==0.20.8
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==1.0.2
pycocotools==2.0
pycryptodomex==3.20.0
Pygments==2.12.0
pymongo==4.6.1
PyMySQL==1.0.2
pyparsing==3.0.9
pyppmd==1.1.0
pyproj==3.3.1
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
pyzstd==0.15.9
qudida==0.0.4
rarfile==4.1
regex==2023.10.3
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3986==1.5.0
rsa==4.9
s3transfer==0.10.0
safetensors==0.3.3
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
sentencepiece==0.1.97
sentry-sdk==1.10.1
setproctitle==1.3.2
Shapely==1.8.2
shortuuid==1.0.9
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.11.0
soupsieve==2.3.2.post1
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.3.0
starlette==0.37.0
strawberry-graphql==0.138.1
submitit==1.5.1
tabulate==0.9.0
taskgroup==0.0.0a4
tbb==2021.6.0
tenacity==8.2.3
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.4.0
terminado==0.15.0
texttable==1.7.0
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.9.7
tinycss2==1.1.1
tokenizers==0.11.2
tomli==2.0.1
torch==1.12.0+cu113
torchaudio==0.12.0+cu113
torchmetrics==1.3.1
torchvision==0.13.0+cu113
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.20.1
tsnecuda==3.0.1+cu113
typing_extensions==4.3.0
tzlocal==5.2
unicode==2.9
Unidecode==1.3.4
unidic==1.1.0
unidic-lite==1.0.8
universal-analytics-python3==1.1.1
urllib3==1.26.12
vitaldb==1.3.3
voxel51-eta==0.12.4
wandb==0.13.4
wasabi==0.10.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.3.7
wfdb==4.1.0
widgetsnbextension==3.6.1
wrapt==1.16.0
wsproto==1.2.0
xmltodict==0.13.0
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.1
yarl==1.9.4
zipp==3.8.0
